Introduce option to keep original files
By setting ALLOW_ORIGINAL_DOWNLOAD to 1, the original files will not be
deleted after creating the scaled versions. This is the foundation for
making the full size images downloadable.
By default, ALLOW_ORIGINAL_DOWNLOAD is 0 and original images will be
deleted while Jekyll does its thing.

+++ _plugins/destroy_originals.rb
@@ -9,14 +9,17 @@ module Jekyll
     end
 
     def self.destroy
-      directories = jekyll_config["image_processing"].each_with_object([]) do |(size, size_options), array|
-        directory = File.join(jekyll_config["destination"], size_options["source"])
-        if directory != jekyll_config["destination"]
-          array.push(directory)
+      # remove the original files when downloads are disabled
+      unless jekyll_config["env"]["ALLOW_ORIGINAL_DOWNLOAD"] == "1"
+        directories = jekyll_config["image_processing"].each_with_object([]) do |(size, size_options), array|
+          directory = File.join(jekyll_config["destination"], size_options["source"])
+          if directory != jekyll_config["destination"]
+            array.push(directory)
+          end
         end
-      end
-      directories.each do |directory|
-        FileUtils.rm_f(Dir.glob("#{directory}/*.[jJ][pP]*[gG]"))
+        directories.each do |directory|
+          FileUtils.rm_f(Dir.glob("#{directory}/*.[jJ][pP]*[gG]"))
+        end
       end
     end
   end
